compiler warnings in osx

+++ modules/ruby/ruby_bound_method.cpp
@@ -1,5 +1,5 @@
 /**
- * Appcelerator Titanium - licensed under the Apache Public License 2
+ * Appcelerator Kroll - licensed under the Apache Public License 2
  * see LICENSE in the root folder for details on the license.
  * Copyright (c) 2008 Appcelerator, Inc. All Rights Reserved.
  */
@@ -20,7 +20,7 @@ namespace kroll {
 	SharedValue RubyBoundMethod::Call(const ValueList& args)
 	{
 		VALUE* ruby_args = (VALUE*)malloc(sizeof(VALUE)*args.size()+1);
-		for (int i = 0; i < args.size(); i++) {
+		for (size_t i = 0; i < args.size(); i++) {
 			ruby_args[i] = RubyUtils::ToRubyValue(args[i]);
 		}
